+++ parser/sql/dialect/hive/src/main/java/org/apache/shardingsphere/sql/parser/hive/visitor/statement/type/HiveDDLStatementVisitor.java
@@ -35,6 +35,8 @@ import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.TableNam
 import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.TruncateTableContext;
 import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.AlterTableContext;
 import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.MsckStatementContext;
+import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.ChangeColumnContext;
+import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.ColumnNameContext;
 import org.apache.shardingsphere.sql.parser.hive.visitor.statement.HiveStatementVisitor;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.column.ColumnDefinitionSegment;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.constraint.ConstraintDefinitionSegment;
@@ -52,6 +54,13 @@ import org.apache.shardingsphere.sql.parser.statement.core.statement.type.ddl.ta
 import org.apache.shardingsphere.sql.parser.statement.core.value.collection.CollectionValue;
 import org.apache.shardingsphere.sql.parser.statement.core.value.identifier.IdentifierValue;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.type.ddl.TruncateStatement;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.column.alter.ChangeColumnDefinitionSegment;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.column.position.ColumnFirstPositionSegment;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.column.position.ColumnAfterPositionSegment;
+import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.AddColumnsContext;
+import org.apache.shardingsphere.sql.parser.autogen.HiveStatementParser.ReplaceColumnsContext;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.column.alter.AddColumnDefinitionSegment;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.column.alter.ReplaceColumnDefinitionSegment;
 import java.util.Collections;
 
 /**
@@ -82,6 +91,18 @@ public final class HiveDDLStatementVisitor extends HiveStatementVisitor implemen
     public ASTNode visitAlterTable(final AlterTableContext ctx) {
         AlterTableStatement result = new AlterTableStatement(getDatabaseType());
         result.setTable((SimpleTableSegment) visit(ctx.alterTableCommonClause().tableName()));
+        if (null != ctx.changeColumn()) {
+            ChangeColumnDefinitionSegment changeColumnSegment = (ChangeColumnDefinitionSegment) visit(ctx.changeColumn());
+            result.getChangeColumnDefinitions().add(changeColumnSegment);
+        }
+        if (null != ctx.addColumns()) {
+            AddColumnDefinitionSegment addSeg = (AddColumnDefinitionSegment) visit(ctx.addColumns());
+            result.getAddColumnDefinitions().add(addSeg);
+        }
+        if (null != ctx.replaceColumns()) {
+            ReplaceColumnDefinitionSegment repSeg = (ReplaceColumnDefinitionSegment) visit(ctx.replaceColumns());
+            result.getReplaceColumnDefinitions().add(repSeg);
+        }
         if (null != ctx.COMPACT()) {
             String compactionType = ctx.string_().getText().replace("'", "");
             if (!isValidCompactionType(compactionType)) {
@@ -160,6 +181,48 @@ public final class HiveDDLStatementVisitor extends HiveStatementVisitor implemen
                 new IdentifierValue(ctx.columnName().getText()));
         DataTypeSegment dataType = (DataTypeSegment) visit(ctx.dataTypeClause());
         return new ColumnDefinitionSegment(ctx.getStart().getStartIndex(), ctx.getStop().getStopIndex(), column, dataType, false, false, getText(ctx));
+    }
+    
+    @Override
+    public ASTNode visitChangeColumn(final ChangeColumnContext ctx) {
+        ColumnSegment oldColumn = new ColumnSegment(ctx.columnName(0).getStart().getStartIndex(), ctx.columnName(0).getStop().getStopIndex(),
+                new IdentifierValue(ctx.columnName(0).getText()));
+        ColumnSegment newColumn = new ColumnSegment(ctx.columnName(1).getStart().getStartIndex(), ctx.columnName(1).getStop().getStopIndex(),
+                new IdentifierValue(ctx.columnName(1).getText()));
+        DataTypeSegment dataType = (DataTypeSegment) visit(ctx.dataTypeClause());
+        ColumnDefinitionSegment columnDefinition = new ColumnDefinitionSegment(ctx.columnName(1).getStart().getStartIndex(),
+                ctx.dataTypeClause().getStop().getStopIndex(), newColumn, dataType, false, false, getText(ctx));
+        ChangeColumnDefinitionSegment result = new ChangeColumnDefinitionSegment(ctx.getStart().getStartIndex(), ctx.getStop().getStopIndex(), columnDefinition);
+        result.setPreviousColumn(oldColumn);
+        if (null != ctx.FIRST()) {
+            ColumnFirstPositionSegment firstPos = new ColumnFirstPositionSegment(ctx.FIRST().getSymbol().getStartIndex(), ctx.FIRST().getSymbol().getStopIndex(), null);
+            result.setColumnPosition(firstPos);
+        } else if (null != ctx.AFTER()) {
+            ColumnNameContext afterCtx = ctx.columnName(ctx.columnName().size() - 1);
+            ColumnSegment afterColumn = new ColumnSegment(afterCtx.getStart().getStartIndex(), afterCtx.getStop().getStopIndex(),
+                    new IdentifierValue(afterCtx.getText()));
+            ColumnAfterPositionSegment afterPos = new ColumnAfterPositionSegment(afterCtx.getStart().getStartIndex(), afterCtx.getStop().getStopIndex(), afterColumn);
+            result.setColumnPosition(afterPos);
+        }
+        return result;
+    }
+    
+    @Override
+    public ASTNode visitAddColumns(final AddColumnsContext ctx) {
+        java.util.Collection<ColumnDefinitionSegment> cols = new java.util.LinkedList<>();
+        for (ColumnDefinitionContext each : ctx.columnDefinition()) {
+            cols.add((ColumnDefinitionSegment) visit(each));
+        }
+        return new AddColumnDefinitionSegment(ctx.getStart().getStartIndex(), ctx.getStop().getStopIndex(), cols);
+    }
+    
+    @Override
+    public ASTNode visitReplaceColumns(final ReplaceColumnsContext ctx) {
+        java.util.Collection<ColumnDefinitionSegment> cols = new java.util.LinkedList<>();
+        for (ColumnDefinitionContext each : ctx.columnDefinition()) {
+            cols.add((ColumnDefinitionSegment) visit(each));
+        }
+        return new ReplaceColumnDefinitionSegment(ctx.getStart().getStartIndex(), ctx.getStop().getStopIndex(), cols);
     }
     
     private String getText(final ParserRuleContext ctx) {